+++ authority-portal-backend/authority-portal-db/src/main/resources/db/migration/V6__AP_Release_1_0_0.sql
@@ -13,3 +13,27 @@ alter table "user"
 drop type user_registration_status;
 
 alter type tmp_enum rename to user_registration_status;
+
+-- Component & Connector online status
+create type component_type as enum ('BROKER', 'DAPS', 'LOGGING_HOUSE');
+create type component_online_status as enum ('UP', 'DOWN', 'PENDING', 'MAINTENANCE');
+create table "component_downtimes"
+(
+    component   component_type           not null,
+    status      component_online_status  not null,
+    environment text                     not null,
+    time_stamp  timestamp with time zone not null,
+    primary key (component, environment, time_stamp)
+);
+create index component_downtimes_time_stamp_index on "component_downtimes" (time_stamp);
+
+create type connector_uptime_status as enum ('UP', 'DOWN', 'DEAD');
+create table "connector_downtimes"
+(
+    connector_id text                     not null,
+    status       connector_uptime_status  not null,
+    environment  text                     not null,
+    time_stamp   timestamp with time zone not null,
+    primary key (connector_id, time_stamp)
+);
+create index connector_downtimes_time_stamp_index on "connector_downtimes" (time_stamp);